minor heading rephrasing
I propose we change the heading to be consistent with the other headings:

As an IntelliJ user, I was looking for the "IntelliJ" header.  

All the other headings start with the name of the IDE (e.g., Atom, Emacs, Vim, etc.).  The IntelliJ section is the only one that puts the name of the plugin rather than the name of the IDE.