change string example from doseq to let
The existing example uses a string as a sequence, and thus prints one character at a time.
This change swaps out doseq for let and updates the binding name to make the example a little bit more practical, while maintaining the spirit of the example (nested calls being captured by with-out-str).